refactor: split the CMS post editor into focused modules
