+++ cms/src/components/PostEditor.tsx
@@ -5,113 +5,30 @@
  * Supports Cmd+S save, new category detection, and unsaved changes warning.
  */
 
-import { BlockNoteSchema, createCodeBlockSpec, defaultBlockSpecs } from '@blocknote/core';
 import { useCreateBlockNote } from '@blocknote/react';
-import { BlockNoteView } from '@blocknote/shadcn';
 import '@blocknote/shadcn/style.css';
 import { Icon } from '@iconify/react';
-import { type KeyboardEvent as ReactKeyboardEvent, useCallback, useEffect, useRef, useState } from 'react';
-import { ErrorBoundary, type FallbackProps } from 'react-error-boundary';
+import { useCallback, useEffect, useRef, useState } from 'react';
 import { toast } from 'sonner';
 import { CategoryMappingDialog } from '@/components/CategoryMappingDialog';
-import { EditorTOC } from '@/components/EditorTOC';
-import { FrontmatterEditor, type FrontmatterEditorRef } from '@/components/FrontmatterEditor';
-import { MarkdownPreview } from '@/components/MarkdownPreview';
+import type { FrontmatterEditorRef } from '@/components/FrontmatterEditor';
+import { blocksToMarkdown, markdownToBlocks, postEditorSchema } from '@/components/post-editor/editor';
+import { PostEditorCanvas } from '@/components/post-editor/PostEditorCanvas';
+import { PostEditorHeader } from '@/components/post-editor/PostEditorHeader';
+import { PostEditorSidebar, type SidebarTab } from '@/components/post-editor/PostEditorSidebar';
+import { SidebarResizeHandle } from '@/components/post-editor/SidebarResizeHandle';
+import { useSidebarResize } from '@/components/post-editor/useSidebarResize';
 import { Button } from '@/components/ui/button';
 import { useEditorHeadings } from '@/hooks';
 import { readPost, writePost } from '@/lib/api';
 import { detectNewCategories, getCategoryMap, setCategoryMap } from '@/lib/category';
 import { DEV_SERVER_URL } from '@/lib/config';
-import { cn } from '@/lib/utils';
 import type { BlogSchema } from '@/types';
-
-// Supported languages for code blocks
-const CODE_BLOCK_LANGUAGES = {
-  typescript: { name: 'TypeScript', aliases: ['ts'] },
-  javascript: { name: 'JavaScript', aliases: ['js'] },
-  tsx: { name: 'TSX' },
-  jsx: { name: 'JSX' },
-  html: { name: 'HTML' },
-  css: { name: 'CSS' },
-  json: { name: 'JSON' },
-  yaml: { name: 'YAML', aliases: ['yml'] },
-  markdown: { name: 'Markdown', aliases: ['md'] },
-  bash: { name: 'Bash', aliases: ['sh', 'shell'] },
-  python: { name: 'Python', aliases: ['py'] },
-  go: { name: 'Go' },
-  rust: { name: 'Rust', aliases: ['rs'] },
-  sql: { name: 'SQL' },
-  c: { name: 'C' },
-  cpp: { name: 'C++', aliases: ['c++'] },
-  java: { name: 'Java' },
-  php: { name: 'PHP' },
-  ruby: { name: 'Ruby', aliases: ['rb'] },
-  swift: { name: 'Swift' },
-  kotlin: { name: 'Kotlin', aliases: ['kt'] },
-  text: { name: 'Plain Text' },
-};
-
-// Create schema with built-in code block using predefined languages
-const schema = BlockNoteSchema.create({
-  blockSpecs: {
-    ...defaultBlockSpecs, // Keep all default blocks (paragraph, heading, list, etc.)
-    codeBlock: createCodeBlockSpec({
-      indentLineWithTab: true,
-      defaultLanguage: 'text',
-      supportedLanguages: CODE_BLOCK_LANGUAGES,
-    }),
-  },
-});
 
 interface PostEditorProps {
   postId: string;
   onClose: () => void;
   onSaved?: () => void;
-}
-
-/**
- * Fallback component for editor errors
- */
-function EditorErrorFallback({ error, resetErrorBoundary }: FallbackProps) {
-  return (
-    <div className="flex h-full flex-col items-center justify-center gap-4 p-8 text-center">
-      <Icon icon="ri:error-warning-line" className="size-12 text-destructive" />
-      <div>
-        <h3 className="font-semibold text-lg">Editor failed to load</h3>
-        <p className="mt-1 text-muted-foreground text-sm">{error.message}</p>
-      </div>
-      <Button variant="outline" onClick={resetErrorBoundary}>
-        Try Again
-      </Button>
-    </div>
-  );
-}
-
-/**
- * Converts BlockNote blocks to markdown
- */
-async function blocksToMarkdown(editor: ReturnType<typeof useCreateBlockNote>): Promise<string> {
-  return await editor.blocksToMarkdownLossy(editor.document);
-}
-
-/**
- * Converts markdown to BlockNote blocks
- */
-async function markdownToBlocks(editor: ReturnType<typeof useCreateBlockNote>, markdown: string): Promise<void> {
-  const blocks = await editor.tryParseMarkdownToBlocks(markdown);
-  editor.replaceBlocks(editor.document, blocks);
-}
-
-type SidebarTab = 'frontmatter' | 'toc' | 'preview';
-
-const SIDEBAR_WIDTH_KEY = 'cms-sidebar-width';
-const SIDEBAR_MIN_WIDTH = 240;
-const SIDEBAR_DEFAULT_WIDTH = 320;
-const SIDEBAR_MAX_WIDTH = 640;
-const SIDEBAR_KEYBOARD_STEP = 16;
-
-function clampSidebarWidth(width: number): number {
-  return Math.min(SIDEBAR_MAX_WIDTH, Math.max(SIDEBAR_MIN_WIDTH, width));
 }
 
 export function PostEditor({ postId, onClose, onSaved }: PostEditorProps) {
@@ -122,32 +39,22 @@ export function PostEditor({ postId, onClose, onSaved }: PostEditorProps) {
   const [showSidebar, setShowSidebar] = useState(true);
   const [sidebarTab, setSidebarTab] = useState<SidebarTab>('frontmatter');
 
-  // Sidebar resize state
-  const [sidebarWidth, setSidebarWidth] = useState(() => {
-    const saved = localStorage.getItem(SIDEBAR_WIDTH_KEY);
-    const savedWidth = saved ? Number(saved) : SIDEBAR_DEFAULT_WIDTH;
-    return Number.isFinite(savedWidth) ? clampSidebarWidth(savedWidth) : SIDEBAR_DEFAULT_WIDTH;
-  });
-  const [isResizing, setIsResizing] = useState(false);
+  const { sidebarWidth, isResizing, startResizing, handleResizeKeyDown } = useSidebarResize();
 
-  // Frontmatter state
   const [frontmatter, setFrontmatter] = useState<BlogSchema>({ title: '' });
   const frontmatterRef = useRef<FrontmatterEditorRef>(null);
 
-  // Preview state
   const [previewContent, setPreviewContent] = useState('');
 
-  // Category state
   const [currentCategories, setCurrentCategories] = useState<string[]>([]);
   const [pendingCategoryMappings, setPendingCategoryMappings] = useState<Record<string, string>>({});
   const [showCategoryDialog, setShowCategoryDialog] = useState(false);
 
   // BlockNote editor with code block language support
-  const editor = useCreateBlockNote({ schema });
+  const editor = useCreateBlockNote({ schema: postEditorSchema });
   const initialContentLoaded = useRef(false);
   const initialFrontmatterLoaded = useRef(false);
 
-  // Extract headings for TOC
   const headings = useEditorHeadings(editor);
 
   // Load post data
@@ -280,54 +187,6 @@ export function PostEditor({ postId, onClose, onSaved }: PostEditorProps) {
     return () => window.removeEventListener('keydown', handleKeyDown);
   }, [handleSave]);
 
-  // Sidebar resize handling
-  useEffect(() => {
-    if (!isResizing) return;
-
-    const handleMouseMove = (e: MouseEvent) => {
-      const newWidth = window.innerWidth - e.clientX;
-      setSidebarWidth(clampSidebarWidth(newWidth));
-    };
-
-    const handleMouseUp = () => {
-      setIsResizing(false);
-      // Persist width to localStorage
-      localStorage.setItem(SIDEBAR_WIDTH_KEY, String(sidebarWidth));
-    };
-
-    document.addEventListener('mousemove', handleMouseMove);
-    document.addEventListener('mouseup', handleMouseUp);
-
-    // Prevent text selection during drag
-    document.body.style.userSelect = 'none';
-    document.body.style.cursor = 'col-resize';
-
-    return () => {
-      document.removeEventListener('mousemove', handleMouseMove);
-      document.removeEventListener('mouseup', handleMouseUp);
-      document.body.style.userSelect = '';
-      document.body.style.cursor = '';
-    };
-  }, [isResizing, sidebarWidth]);
-
-  const handleSidebarResizeKeyDown = useCallback(
-    (event: ReactKeyboardEvent<HTMLHRElement>) => {
-      let nextWidth: number | null = null;
-
-      if (event.key === 'ArrowLeft') nextWidth = sidebarWidth + SIDEBAR_KEYBOARD_STEP;
-      if (event.key === 'ArrowRight') nextWidth = sidebarWidth - SIDEBAR_KEYBOARD_STEP;
-      if (event.key === 'Home') nextWidth = SIDEBAR_MIN_WIDTH;
-      if (event.key === 'End') nextWidth = SIDEBAR_MAX_WIDTH;
-      if (nextWidth === null) return;
-
-      event.preventDefault();
-      const clampedWidth = clampSidebarWidth(nextWidth);
-      setSidebarWidth(clampedWidth);
-      localStorage.setItem(SIDEBAR_WIDTH_KEY, String(clampedWidth));
-    },
-    [sidebarWidth],
-  );
-
   // Warn about unsaved changes
   useEffect(() => {
     const handleBeforeUnload = (e: BeforeUnloadEvent) => {
@@ -350,17 +209,13 @@ export function PostEditor({ postId, onClose, onSaved }: PostEditorProps) {
     onClose();
   }, [hasUnsavedChanges, onClose]);
 
-  // Get post preview URL (points to Astro dev server)
-  const getPreviewUrl = () => {
-    // Use frontmatter.link if available, otherwise extract filename from postId
-    const slug =
-      frontmatter.link ||
-      postId
-        .replace(/\.mdx?$/, '')
-        .split('/')
-        .pop();
-    return `${DEV_SERVER_URL}/post/${slug}`;
-  };
+  const previewSlug =
+    frontmatter.link ||
+    postId
+      .replace(/\.mdx?$/, '')
+      .split('/')
+      .pop();
+  const previewUrl = `${DEV_SERVER_URL}/post/${previewSlug}`;
 
   // Handle TOC navigation
   const handleTOCNavigate = useCallback(
@@ -423,164 +278,43 @@ export function PostEditor({ postId, onClose, onSaved }: PostEditorProps) {
 
   return (
     <div className="fixed inset-0 z-50 flex flex-col bg-background">
-      {/* Header */}
-      <header className="flex items-center justify-between border-border border-b px-4 py-2">
-        <div className="flex items-center gap-3">
-          <button
-            type="button"
-            onClick={handleClose}
-            className="rounded-lg p-2 text-muted-foreground transition-colors hover:bg-muted hover:text-foreground"
-            title="Close editor"
-          >
-            <Icon icon="ri:arrow-left-line" className="size-5" />
-          </button>
-          <div>
-            <h1 className="line-clamp-1 font-medium">{frontmatter.title || 'Untitled'}</h1>
-            <p className="text-muted-foreground text-xs">{postId}</p>
-          </div>
-          {hasUnsavedChanges && <span className="rounded bg-orange-500/10 px-2 py-0.5 text-orange-500 text-xs">Unsaved</span>}
-        </div>
+      <PostEditorHeader
+        title={frontmatter.title}
+        postId={postId}
+        previewUrl={previewUrl}
+        hasUnsavedChanges={hasUnsavedChanges}
+        isSaving={isSaving}
+        showSidebar={showSidebar}
+        onClose={handleClose}
+        onSave={handleSave}
+        onToggleSidebar={() => setShowSidebar(!showSidebar)}
+      />
 
-        <div className="flex items-center gap-2">
-          {/* Preview link */}
-          <a
-            href={getPreviewUrl()}
-            target="_blank"
-            rel="noopener noreferrer"
-            className="flex items-center gap-1 rounded-lg px-3 py-2 text-muted-foreground text-sm transition-colors hover:bg-muted hover:text-foreground"
-          >
-            <Icon icon="ri:external-link-line" className="size-4" />
-            Preview
-          </a>
-
-          {/* Toggle sidebar */}
-          <button
-            type="button"
-            onClick={() => setShowSidebar(!showSidebar)}
-            className={cn(
-              'rounded-lg p-2 transition-colors',
-              showSidebar ? 'bg-muted text-foreground' : 'text-muted-foreground hover:bg-muted',
-            )}
-            title="Toggle frontmatter panel"
-          >
-            <Icon icon="ri:sidebar-unfold-line" className="size-5" />
-          </button>
-
-          {/* Save button */}
-          <Button onClick={handleSave} disabled={isSaving || !hasUnsavedChanges}>
-            {isSaving ? (
-              <>
-                <Icon icon="ri:loader-4-line" className="mr-1.5 size-4 animate-spin" />
-                Saving...
-              </>
-            ) : (
-              <>
-                <Icon icon="ri:save-line" className="mr-1.5 size-4" />
-                Save
-              </>
-            )}
-          </Button>
-        </div>
-      </header>
-
-      {/* Main content */}
       <div className="flex flex-1 overflow-hidden">
-        {/* Editor */}
-        <main className="flex-1 overflow-auto">
-          <div className="mx-auto max-w-3xl p-6">
-            <ErrorBoundary FallbackComponent={EditorErrorFallback}>
-              <BlockNoteView editor={editor} theme="dark" />
-            </ErrorBoundary>
-          </div>
-        </main>
+        <PostEditorCanvas editor={editor} />
 
-        {/* Resize Handle */}
         {showSidebar && (
-          <hr
-            tabIndex={0}
-            aria-orientation="vertical"
-            aria-label="Resize sidebar"
-            aria-valuemin={SIDEBAR_MIN_WIDTH}
-            aria-valuemax={SIDEBAR_MAX_WIDTH}
-            aria-valuenow={sidebarWidth}
-            aria-valuetext={`${sidebarWidth} pixels`}
-            className={cn(
-              'h-full w-1 shrink-0 cursor-col-resize border-none transition-colors hover:bg-primary/50',
-              isResizing && 'bg-primary',
-            )}
-            onMouseDownCapture={() => setIsResizing(true)}
-            onKeyDownCapture={handleSidebarResizeKeyDown}
+          <SidebarResizeHandle
+            width={sidebarWidth}
+            isResizing={isResizing}
+            onMouseDown={startResizing}
+            onKeyDown={handleResizeKeyDown}
           />
         )}
 
-        {/* Sidebar */}
         {showSidebar && (
-          <aside style={{ width: sidebarWidth }} className="flex shrink-0 flex-col border-border border-l bg-card">
-            {/* Tab buttons */}
-            <div className="flex border-border border-b">
-              <button
-                type="button"
-                onClick={() => handleTabChange('frontmatter')}
-                className={cn(
-                  'flex-1 px-3 py-2.5 font-medium text-sm transition-colors',
-                  sidebarTab === 'frontmatter'
-                    ? 'border-primary border-b-2 text-foreground'
-                    : 'text-muted-foreground hover:text-foreground',
-                )}
-              >
-                <Icon icon="ri:settings-3-line" className="mr-1 inline-block size-4" />
-                属性
-              </button>
-              <button
-                type="button"
-                onClick={() => handleTabChange('toc')}
-                className={cn(
-                  'flex-1 px-3 py-2.5 font-medium text-sm transition-colors',
-                  sidebarTab === 'toc'
-                    ? 'border-primary border-b-2 text-foreground'
-                    : 'text-muted-foreground hover:text-foreground',
-                )}
-              >
-                <Icon icon="ri:list-unordered" className="mr-1 inline-block size-4" />
-                目录
-              </button>
-              <button
-                type="button"
-                onClick={() => handleTabChange('preview')}
-                className={cn(
-                  'flex-1 px-3 py-2.5 font-medium text-sm transition-colors',
-                  sidebarTab === 'preview'
-                    ? 'border-primary border-b-2 text-foreground'
-                    : 'text-muted-foreground hover:text-foreground',
-                )}
-              >
-                <Icon icon="ri:eye-line" className="mr-1 inline-block size-4" />
-                预览
-              </button>
-            </div>
-
-            {/* Tab content */}
-            <div className="flex-1 overflow-auto">
-              {sidebarTab === 'frontmatter' && (
-                <FrontmatterEditor
-                  ref={frontmatterRef}
-                  frontmatter={frontmatter}
-                  onChange={handleFrontmatterChange}
-                  onCategoriesChange={handleCategoriesChange}
-                />
-              )}
-              {sidebarTab === 'toc' && (
-                <div className="p-4">
-                  <EditorTOC headings={headings} onNavigate={handleTOCNavigate} />
-                </div>
-              )}
-              {sidebarTab === 'preview' && (
-                <div className="p-4">
-                  <MarkdownPreview content={previewContent} />
-                </div>
-              )}
-            </div>
-          </aside>
+          <PostEditorSidebar
+            width={sidebarWidth}
+            activeTab={sidebarTab}
+            frontmatter={frontmatter}
+            frontmatterRef={frontmatterRef}
+            headings={headings}
+            previewContent={previewContent}
+            onTabChange={handleTabChange}
+            onFrontmatterChange={handleFrontmatterChange}
+            onCategoriesChange={handleCategoriesChange}
+            onTOCNavigate={handleTOCNavigate}
+          />
         )}
       </div>
